server/python/requirements.txt
insightface
deepface
faiss-cpu
onnxruntime
opencv-python-headless
numpy
requests
pillow
